make atomicGraphs by factory pattern
